refactor: update the sidebar toggle button file name
